move get_dummy_args to util file
